clean up projectts pill bar at hhome drawer
